+++ e-library/src/Components/images/images.jsx
@@ -1,14 +1,12 @@
 import React from 'react';
 
-function Images() {
-  return (
-    <div className="img">
-      <img
-        src="https://res.cloudinary.com/silasmanny/image/upload/v1650552036/q0bajuhzmyvdx5jplj1p.jpg"
-        alt="In your own backyard"
-      />
-    </div>
-  );
-}
+const Images = () => (
+  <div className="img">
+    <img
+      src="https://res.cloudinary.com/silasmanny/image/upload/v1650552036/q0bajuhzmyvdx5jplj1p.jpg"
+      alt="In your own backyard"
+    />
+  </div>
+);
 
 export default Images;
